requirements.txt
anthropic>=0.39.0
python-docx>=1.1.0
click>=8.1.0
PyPDF2>=3.0.0
tenacity>=8.0.0
